deps: Split dependencies.base
Split the Factory and dependency classes out
of the base.py script to improve maintainability.
