+++ mesonbuild/dependencies/hdf5.py
@@ -22,10 +22,10 @@ import subprocess
 from pathlib import Path
 
 from ..mesonlib import OrderedSet, join_args
-from .base import (
-    DependencyException, DependencyMethods, ConfigToolDependency,
-    PkgConfigDependency, factory_methods
-)
+from .base import DependencyException, DependencyMethods
+from .configtool import ConfigToolDependency
+from .pkgconfig import PkgConfigDependency
+from .factory import factory_methods
 import typing as T
 
 if T.TYPE_CHECKING:
